Reapier bug with netlify auth, Content Wrapper make more visible
