+++ src/pages/index.js
@@ -6,26 +6,24 @@ import ContentWrapper from "../components/ContentWrapper"
 // import AboutChallenge from "../components/AboutChallenge"
 import { loadState } from '../localStorage'
 import FirstTimeUser from "../components/FirstTimeOnApp/FirstTimeUser"
-
+const setHandle =()=>{
+  if(loadState('listOfResonsArray')){
+   return(false)
+  }else{
+    return(true)
+  }
+}
 const IndexPage = () => {
-  const [firstTime, setFirstTime] = React.useState(loadState('goalItem'))
-   
-  const setHandle =()=>{
-    if(loadState('listOfResonsArray')){
-      setFirstTime(true)
-    }else{
-      setFirstTime(false)
-    }
-    setHandle()
+  const [firstTime, setFirstTime] = React.useState(setHandle())
+  const setHandleFalse =()=>{
+    setFirstTime(false)
   }
 return(
   <Layout>
     <SEO title="Home" />
-       {/* {firstTime ? <FirstTimeUser /> : <>
+       {firstTime ? <FirstTimeUser setStart={setHandleFalse} /> : <>
        <MenuPanel /><ContentWrapper />
-       </>} */}
-       <MenuPanel /><ContentWrapper />
-
+       </>}
   </Layout>
   )
 }
