+++ src/pages/login.js
@@ -1,28 +1,27 @@
-import React, { Component } from "react"
+import React, { useState, useEffect } from "react"
 import { Link } from "gatsby"
 import Layout from "../components/layout"
 import Image from "../components/image"
 import SEO from "../components/seo"
 import { loadState } from "../localStorage"
 import { getFirebase } from "../../firebase"
+
 function initNetlifyIdentity() {
   console.log("initNetlifyIdentity called.")
   const script = document.createElement("script");
-
   script.src = "https://identity.netlify.com/v1/netlify-identity-widget.js"
   script.async = true;
-  
   document.body.appendChild(script);
 }
 let netlifyIdentity
 
-class NetlifyIdentity extends Component {
-  componentDidMount() {
+//////
+const NetlifyIdentity = () => {
+  useEffect(() => {
     netlifyIdentity = window.netlifyIdentity;
     initNetlifyIdentity();
     const lazyApp = import('firebase/app')
     const lazyDatabase = import('firebase/database')
-  
     Promise.all([lazyApp, lazyDatabase]).then(([firebase]) => {
       const database = getFirebase(firebase).database()
       let data = {
@@ -30,16 +29,7 @@ class NetlifyIdentity extends Component {
       }
       database.ref().set(data);
     })
-  }
-
-  
-
-  render() {
-    
-    return(<div></div>)
-  }
-}
-const IndexPage = () => {
+  });
   function openNetlifyModal() {
     if(netlifyIdentity){
       netlifyIdentity.open();
@@ -47,15 +37,15 @@ const IndexPage = () => {
     else
       console.log('netlifyIdentity not defined')
   }
+
+
   return(
     <Layout>
-      <NetlifyIdentity />
       <SEO title="Home" />
       <h1>Hi people</h1>
-      <h2 onClick={() => { openNetlifyModal() }}>Login</h2>
+      <h2 onClick={openNetlifyModal}>Login</h2>
       <p>Welcome to your new Gatsby site.</p>
-
-      <p>Now go build something great.</p>{console.log(loadState('gotrue.user'))}
+      <p>Now go build something great.</p>{console.log("asdad" ,loadState('gotrue.user'))}
       <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
         <Image />
       </div>
@@ -63,5 +53,4 @@ const IndexPage = () => {
     </Layout> 
   )
 }
-
-export default IndexPage
+export default NetlifyIdentity
